Remove simulation helpers from Data Distributor
